* AtomicChange goes to change.py * fix KeyError's in Namespace * clear _dirty after commit
